# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "starker Go API Template",
        "title": "starker Go API Template",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8000",
    "basePath": "/api",
    "paths": {
...
        "/teams/{teamID}/projects/{projectID}/services": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new service with Docker compose configuration within a team and project",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "service"
                ],
                "summary": "Create a new service",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Project ID",
                        "name": "projectID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Service creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/servicesvc.CreateServiceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Service"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body, team access denied, or project not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/teams/{teamID}/servers/{serverID}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Gets a specific server configuration by ID within a team that the user has access to",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Get a specific server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "serverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Server"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Team access denied",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates an existing server configuration within a team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Update a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "serverID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Server update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/serversvc.UpdateServerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server updated successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Server"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or team access denied",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a specific server configuration by ID within a team",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "server"
                ],
                "summary": "Delete a server",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Team ID",
                        "name": "teamID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Server ID",
                        "name": "serverID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Server deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/response.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Team access denied",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "User not authenticated",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Server not found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "models.Server": {
            "type": "object",
            "properties": {
                "created_at": {
                    "description": "Timestamp when the server was created",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "description": {
                    "description": "Server description",
                    "type": "string",
                    "example": "Main production server"
                },
                "id": {
                    "description": "Unique identifier for the server",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "ip": {
                    "description": "Server IP address",
                    "type": "string",
                    "example": "192.168.1.100"
                },
                "name": {
                    "description": "Server name",
                    "type": "string",
                    "example": "Production Server"
                },
                "port": {
                    "description": "SSH port",
                    "type": "string",
                    "example": "22"
                },
                "private_key_id": {
                    "description": "Associated private key ID",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "team_id": {
                    "description": "Associated team ID",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "updated_at": {
                    "description": "Timestamp when the server was last updated",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "user": {
                    "description": "SSH username",
                    "type": "string",
                    "example": "ubuntu"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "container_id": {
                    "description": "Docker container ID",
                    "type": "string",
                    "example": "abc123..."
                },
                "created_at": {
                    "description": "Timestamp when the service was created",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "description": {
                    "description": "Service description",
                    "type": "string",
                    "example": "Main web application service"
                },
                "id": {
                    "description": "Unique identifier for the service",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "last_deployed_at": {
                    "description": "Timestamp when the service was last deployed",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                },
                "name": {
                    "description": "Service name",
                    "type": "string",
                    "example": "web-app"
                },
                "project_id": {
                    "description": "Associated project ID",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "server_id": {
                    "description": "Associated server ID",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "state": {
                    "description": "Service state (running, stopped, etc.)",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.ServiceState"
                        }
                    ],
                    "example": "running"
                },
                "team_id": {
                    "description": "Associated team ID",
                    "type": "string",
                    "example": "01ARZ3NDEKTSV4RRFFQ69G5FAV"
                },
                "type": {
                    "description": "Service type (e.g., docker, compose)",
                    "type": "string",
                    "example": "docker"
                },
                "updated_at": {
                    "description": "Timestamp when the service was last updated",
                    "type": "string",
                    "example": "2023-01-01T12:00:00Z"
                }
            }
        },
        "models.ServiceState": {
            "type": "string",
            "enum": [
                "running",
                "stopped",
                "starting",
                "stopping",
                "restarting"
            ],
            "x-enum-varnames": [
                "ServiceStateRunning",
                "ServiceStateStopped",
                "ServiceStateStarting",
                "ServiceStateStopping",
                "ServiceStateRestarting"
            ]
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "err_code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "response.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "serversvc.UpdateServerRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "ip": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "port": {
                    "type": "string",
                    "maxLength": 5,
                    "minLength": 1
                },
                "private_key_id": {
                    "type": "string"
                },
                "user": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "servicesvc.CreateServiceRequest": {
            "type": "object",
            "required": [
                "compose_file",
                "name",
                "server_id",
                "type"
            ],
            "properties": {
                "compose_file": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "maxLength": 500
                },
                "name": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 3
                },
                "server_id": {
                    "type": "string"
                },
                "type": {
                    "type": "string",
                    "enum": [
                        "docker",
                        "compose"
                    ]
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter JWT Bearer token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}